+++ packages/plugin-spoiler/web/src/editor.ts
@@ -4,7 +4,7 @@ import SpoilerEditorWrapper from './Components/Wrappers/SpoilerEditorWrapper';
 import { EditorPluginCreator } from 'wix-rich-content-common';
 import { DEFAULTS } from './defaults';
 import { createTiptapExtensions } from './tiptap';
-import { TiptapEditorPlugin } from 'wix-tiptap-editor';
+import { TiptapEditorPlugin } from 'ricos-tiptap-types';
 export { SpoilerEditorWrapper };
 export { default as BlockSpoilerComponent } from './Components/BlockSpoilerComponent';