+++ packages/tiptap-editor/web/src/commands/insertNode.ts
@@ -1,0 +1,20 @@
+import { RawCommands } from '@tiptap/core';
+import { generateId } from 'ricos-content';
+
+declare module '@tiptap/core' {
+  interface Commands<ReturnType> {
+    insertNode: {
+      /**
+       * Insert a node.
+       */
+      insertNode: (type: string, attrs: Record<string, unknown>) => ReturnType;
+    };
+  }
+}
+
+export const insertNode: RawCommands['insertNode'] = (type, attrs = {}) => ({ commands }) => {
+  return commands.insertContent({
+    type,
+    attrs: { id: attrs.id || generateId(), ...attrs },
+  });
+};