+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/RicosTiptapEditor.tsx
@@ -1,16 +1,16 @@
 import { Editor, EditorContent, JSONContent } from '@tiptap/react';
+import { Node } from 'prosemirror-model';
 import React, { FunctionComponent, useEffect, useState } from 'react';
-import editorStyles from '../../statics/styles/tiptap-editor-styles.scss';
+import { getLangDir } from 'wix-rich-content-common';
 import { createEditorStyles } from 'wix-rich-content-editor-common';
-import { tiptapToDraft } from '../..';
+import { tiptapExtensions as coreExtensions, tiptapToDraft } from 'wix-tiptap-extensions';
 import { RicosTiptapContext } from '../../context';
 import { useForceUpdate } from '../../lib/useForceUpdate';
 import { Extensions } from '../../models/Extensions';
-import { tiptapExtensions as coreExtensions } from '../../tiptap-extensions';
+import editorStyles from '../../statics/styles/tiptap-editor-styles.scss';
 import { RicosTiptapEditorProps } from '../../types';
 import { coreConfigs } from './core-configs';
-import { getLangDir } from 'wix-rich-content-common';
-import { Node } from 'prosemirror-model';
+// import { patchExtensions } from '../../unsupported-content';
 
 // TODO: maybe should move it to utils ?
 const getSelectedNodes = ({ editor }) => {
@@ -36,14 +36,17 @@ export const RicosTiptapEditor: FunctionComponent<RicosTiptapEditorProps> = ({
 }) => {
   const forceUpdate = useForceUpdate();
   const [editor, setEditor] = useState<Editor>((null as unknown) as Editor);
+  // const patchedExtensions = patchExtensions(content, [...coreConfigs, ...extensions]);
   const mergedExtensions = Extensions.of([...coreConfigs, ...extensions]);
 
   const getContent = editor => tiptapToDraft(editor.getJSON() as JSONContent);
 
   useEffect(() => {
     const tiptapExtensions = mergedExtensions.getTiptapExtensions();
+    const allExtensions = [...coreExtensions, ...tiptapExtensions];
+    console.log({ allExtensions }); // eslint-disable-line no-console
     const editorInstance = new Editor({
-      extensions: [...coreExtensions, ...tiptapExtensions],
+      extensions: allExtensions,
       content,
       injectCSS: true,
       onUpdate: ({ editor }) => {