+++ packages/tiptap-editor/web/src/extensions/extension-bulleted-list/index.ts
@@ -1,5 +1,0 @@
-import { BulletedList } from './bulleted-list';
-
-export * from './bulleted-list';
-
-export const createBulletedList = () => BulletedList;