+++ packages/tiptap-editor/web/src/extensions/extension-color/index.ts
@@ -1,5 +1,0 @@
-import { Color } from './color';
-
-export const createColor = () => {
-  return Color;
-};