+++ packages/ui-components/web/src/Components/SettingsMobileHeader.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ActionButtons } from '..';
+import { Button } from '..';
 import { RichContentTheme, TranslationFunction } from 'wix-rich-content-common';
 
 import styles from '../../statics/styles/settings-mobile-header.scss';
@@ -26,23 +26,22 @@ const SettingsMobileHeader: React.FC<SettingsMobileHeaderProps> = ({
 }) => {
   const saveText = saveLabel || t('SettingsPanelFooter_Save');
   const cancelText = cancelLabel || t('SettingsPanelFooter_Cancel');
+  const buttonsDataHook = { save: 'actionButtonSave', cancel: 'actionButtonCancel' };
+
+  const renderButton = (text, dataHook, onClick) => (
+    <Button ariaLabel={text} theme={theme} isMobile dataHook={dataHook} onClick={onClick}>
+      {text}
+    </Button>
+  );
 
   return (
     <div className={styles.setting_mobile_header}>
-      <ActionButtons
-        isMobile
-        onCancel={onCancel}
-        onSave={onSave}
-        cancelText={cancelText}
-        saveText={saveText}
-        theme={theme}
-        t={t}
-      >
-        <>
-          {title && <div className={styles.setting_mobile_header_title}>{title}</div>}
-          {children}
-        </>
-      </ActionButtons>
+      {renderButton(saveText, buttonsDataHook.save, onSave)}
+      <>
+        {title && <div className={styles.setting_mobile_header_title}>{title}</div>}
+        {children}
+      </>
+      {renderButton(cancelText, buttonsDataHook.cancel, onCancel)}
     </div>
   );
 };